+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - DescriptionListItem: Add `titleElement` prop to pass elements as a title instead of just text
 - Input: Fix element outline missing on iOS Safari
 - Button: Add primary and secondary variants, using "variant" prop
+- Add Alert component